♻️ Create a report collection
