+++ packages/logs/src/boot/startLogs.spec.ts
@@ -1,14 +1,7 @@
-import type { Context, RawReport, TimeStamp } from '@datadog/browser-core'
-import {
-  noop,
-  Observable,
-  resetExperimentalFeatures,
-  updateExperimentalFeatures,
-  stopSessionManager,
-} from '@datadog/browser-core'
+import type { Context, TimeStamp } from '@datadog/browser-core'
+import { noop, stopSessionManager } from '@datadog/browser-core'
 import sinon from 'sinon'
 import { deleteEventBridgeStub, initEventBridgeStub, stubEndpointBuilder } from '../../../core/test/specHelper'
-import { stubReportingObserver } from '../../../core/test/stubReportApis'
 import type { LogsConfiguration } from '../domain/configuration'
 import { validateAndBuildLogsConfiguration } from '../domain/configuration'
 
@@ -50,19 +43,15 @@ describe('logs', () => {
   let baseConfiguration: LogsConfiguration
   let sessionIsTracked: boolean
   let server: sinon.SinonFakeServer
-  let reportObservable: Observable<RawReport>
 
   const sessionManager: LogsSessionManager = {
     findTrackedSession: () => (sessionIsTracked ? { id: SESSION_ID } : undefined),
   }
-  let stopLogs = noop
   const startLogs = ({
-    sender = createSender(noop),
     configuration: configurationOverrides,
   }: { sender?: Sender; configuration?: Partial<LogsConfiguration> } = {}) => {
     const configuration = { ...baseConfiguration, ...configurationOverrides }
-    const startLogs = doStartLogs(configuration, () => undefined, reportObservable, sessionManager, sender)
-    stopLogs = startLogs.stop
+    const startLogs = doStartLogs(configuration, () => undefined, sessionManager)
     return startLogs.send
   }
 
@@ -73,7 +62,6 @@ describe('logs', () => {
       maxBatchSize: 1,
     }
     sessionIsTracked = true
-    reportObservable = new Observable<RawReport>()
     server = sinon.fakeServer.create()
   })
 
@@ -82,7 +70,6 @@ describe('logs', () => {
     delete window.DD_RUM
     deleteEventBridgeStub()
     stopSessionManager()
-    stopLogs()
   })
 
   describe('request', () => {
@@ -152,92 +139,21 @@ describe('logs', () => {
     })
   })
 
-  describe('reports', () => {
-    let sender: Sender
-    let logErrorSpy: jasmine.Spy
-    let reportingObserverStub: ReturnType<typeof stubReportingObserver>
-
-    beforeEach(() => {
-      sender = createSender(noop)
-      logErrorSpy = spyOn(sender, 'sendToHttp')
-      reportingObserverStub = stubReportingObserver()
-    })
-
-    afterEach(() => {
-      reportingObserverStub.reset()
-    })
-
-    it('should send reports when ff forward-reports is enabled', () => {
-      updateExperimentalFeatures(['forward-reports'])
-      const { stop } = originalStartLogs(
-        validateAndBuildLogsConfiguration({ ...initConfiguration, forwardReports: ['intervention'] })!,
-        sender
-      )
-
-      reportingObserverStub.raiseReport('intervention')
-
-      expect(logErrorSpy).toHaveBeenCalled()
-
-      resetExperimentalFeatures()
-      stop()
-    })
-
-    it('should not send reports when ff forward-reports is disabled', () => {
-      const { stop } = originalStartLogs(
-        validateAndBuildLogsConfiguration({ ...initConfiguration, forwardReports: ['intervention'] })!,
-        sender
-      )
-      reportingObserverStub.raiseReport('intervention')
-
-      expect(logErrorSpy).not.toHaveBeenCalled()
-      stop()
-    })
-
-    it('should not send reports when forwardReports init option not specified', () => {
-      const { stop } = originalStartLogs(validateAndBuildLogsConfiguration({ ...initConfiguration })!, sender)
-      reportingObserverStub.raiseReport('intervention')
-
-      expect(logErrorSpy).not.toHaveBeenCalled()
-      stop()
-    })
-
-    it('should add the source file information to the message for non error reports', () => {
-      updateExperimentalFeatures(['forward-reports'])
-      const { stop } = originalStartLogs(
-        validateAndBuildLogsConfiguration({ ...initConfiguration, forwardReports: ['deprecation'] })!,
-        sender
-      )
-
-      reportingObserverStub.raiseReport('deprecation')
-
-      expect(logErrorSpy).toHaveBeenCalledOnceWith(
-        'deprecation: foo bar Found in http://foo.bar/index.js:20:10',
-        undefined,
-        'warn'
-      )
-
-      resetExperimentalFeatures()
-      stop()
-    })
-  })
-
   describe('sampling', () => {
     it('should be applied when event bridge is present', () => {
       const sendSpy = spyOn(initEventBridgeStub(), 'send')
 
       let configuration = { ...baseConfiguration, sampleRate: 0 }
-      let { send, stop } = originalStartLogs(configuration, createSender(noop))
+      let { send } = originalStartLogs(configuration, createSender(noop))
       send(DEFAULT_MESSAGE, {})
 
       expect(sendSpy).not.toHaveBeenCalled()
-      stop()
 
       configuration = { ...baseConfiguration, sampleRate: 100 }
-      ;({ send, stop } = originalStartLogs(configuration, createSender(noop)))
+      ;({ send } = originalStartLogs(configuration, createSender(noop)))
       send(DEFAULT_MESSAGE, {})
 
       expect(sendSpy).toHaveBeenCalled()
-      stop()
     })
   })
